backend/requirements.txt
# Intelli-Credit Decisioning Engine — Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2
pydantic==2.8.2
pydantic-settings==2.4.0

# OCR & Document Processing
pytesseract==0.3.13
Pillow==10.4.0
pdfplumber==0.11.4
pdf2image==1.17.0

# HTML Scraping
beautifulsoup4==4.12.3
lxml==5.3.0

# Data Processing & Graph
polars==1.7.1
networkx==3.2.1
numpy==1.26.4

# SMT Solver (Neuro-Symbolic Verification)
z3-solver==4.13.0.0

# SSE
sse-starlette==2.1.0
